+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.12.0-rc.6
 
+* Rename `nvidia cdi generate` `--root` flag to `--driver-root` to better indicate intent
 * [libnvidia-container] Add nvcubins.bin to DriverStore components under WSL2
 
 ## v1.12.0-rc.5